requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
blas=1.0=openblas
bottleneck=1.3.5=py311ha0d4635_0
brotli=1.0.9=h1a28f6b_7
brotli-bin=1.0.9=h1a28f6b_7
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.08.22=hca03da5_0
contourpy=1.0.5=py311h48ca7d4_0
cycler=0.11.0=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h1192e45_0
giflib=5.2.1=h80987f9_3
gurobi=10.0.3=py311_0
jpeg=9e=h80987f9_1
kiwisolver=1.4.4=py311h313beb8_0
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libbrotlicommon=1.0.9=h1a28f6b_7
libbrotlidec=1.0.9=h1a28f6b_7
libbrotlienc=1.0.9=h1a28f6b_7
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libffi=3.4.4=hca03da5_0
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libtiff=4.5.1=h313beb8_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_0
matplotlib=3.7.2=py311hca03da5_0
matplotlib-base=3.7.2=py311h7aedaa7_0
munkres=1.1.4=py_0
ncurses=6.4=h313beb8_0
numexpr=2.8.7=py311h6dc990b_0
numpy=1.26.0=py311he598dae_0
numpy-base=1.26.0=py311hfbfe69c_0
openjpeg=2.3.0=h7a6adac_2
openssl=1.1.1w=h1a28f6b_0
packaging=23.1=py311hca03da5_0
pandas=2.1.1=py311h7aedaa7_0
pillow=10.0.1=py311h3b245a6_0
pip=23.3=py311hca03da5_0
pyparsing=3.0.9=py311hca03da5_0
python=3.11.0=hc0d8a6c_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2023.3.post1=py311hca03da5_0
readline=8.2=h1a28f6b_0
scipy=1.11.3=py311hc76d9b0_0
setuptools=68.0.0=py311hca03da5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h80987f9_0
tk=8.6.12=hb8d0fd4_0
tornado=6.3.3=py311h80987f9_0
tzdata=2023c=h04d1e81_0
wheel=0.41.2=py311hca03da5_0
xz=5.4.2=h80987f9_0
zlib=1.2.13=h5a0b063_0
zstd=1.5.5=hd90d995_0
